Move CmdLine options into its own file.
Also use CKAN.CmdLine as the namespace.

Otherwise all the code is the same.
