+++ CKAN/CmdLine/Options.cs
@@ -1,0 +1,228 @@
+using System;
+using System.Collections.Generic;
+using CommandLine;
+
+namespace CKAN.CmdLine
+{
+    // Look, parsing options is so easy and beautiful I made
+    // it into a special class for you to admire!
+
+    internal class Options
+    {
+        public Options(string[] args)
+        {
+            if (! Parser.Default.ParseArgumentsStrict(
+                args, new Actions(), (verb, suboptions) =>
+                {
+                action = verb;
+                options = suboptions;
+            }
+            ))
+            {
+                throw (new BadCommandException("Try ckan --help"));
+            }
+
+            // If we're here, success!
+        }
+
+        public string action { get; set; }
+        public object options { get; set; }
+    }
+
+    // Actions supported by our client go here.
+    // TODO: Figure out how to do per action help screens.
+
+    internal class Actions
+    {
+        [VerbOption("gui", HelpText = "Start the CKAN GUI")]
+        public GuiOptions GuiOptions { get; set; }
+
+        [VerbOption("update", HelpText = "Update list of available mods")]
+        public UpdateOptions Update { get; set; }
+
+        [VerbOption("available", HelpText = "List available mods")]
+        public AvailableOptions Available { get; set; }
+
+        [VerbOption("install", HelpText = "Install a KSP mod")]
+        public InstallOptions Install { get; set; }
+
+        [VerbOption("remove", HelpText = "Remove an installed mod")]
+        public RemoveOptions Remove { get; set; }
+
+        [VerbOption("upgrade", HelpText = "Upgrade an installed mod")]
+        public UpgradeOptions Upgrade { get; set; }
+
+        [VerbOption("scan", HelpText = "Scan for manually installed KSP mods")]
+        public ScanOptions Scan { get; set; }
+
+        [VerbOption("list", HelpText = "List installed modules")]
+        public ListOptions List { get; set; }
+
+        [VerbOption("show", HelpText = "Show information about a mod")]
+        public ShowOptions Show { get; set; }
+
+        [VerbOption("clean", HelpText = "Clean away downloaded files from the cache")]
+        public CleanOptions Clean { get; set; }
+
+        [VerbOption("list-installs", HelpText = "List all known KSP installations")]
+        public ListInstallsOptions ListInstalls { get; set; }
+
+        [VerbOption("add-install", HelpText = "Add a new KSP installation")]
+        public AddInstallOptions AddInstall { get; set; }
+
+        [VerbOption("rename-install", HelpText = "Rename a known KSP installation")]
+        public RenameInstallOptions RenameInstall { get; set; }
+
+        [VerbOption("remove-install", HelpText = "Remove a known KSP installation")]
+        public RemoveInstallOptions RemoveInstall { get; set; }
+
+        [VerbOption("set-default-install", HelpText = "Sets a known KSP installation as default")]
+        public SetDefaultInstallOptions SetDefaultInstall { get; set; }
+
+        [VerbOption("clear-cache", HelpText = "Clears the download cache")]
+        public ClearCacheOptions ClearCache { get; set; }
+
+        [VerbOption("version", HelpText = "Show the version of the CKAN client being used.")]
+        public VersionOptions Version { get; set; }
+    }
+
+    // Options common to all classes.
+
+    internal class CommonOptions
+    {
+        [Option('v', "verbose", DefaultValue = false, HelpText = "Show more of what's going on when running.")]
+        public bool Verbose { get; set; }
+
+        [Option('d', "debug", DefaultValue = false, HelpText = "Show debugging level messages. Implies verbose")]
+        public bool Debug { get; set; }
+
+        [Option("ksp", DefaultValue = null, HelpText = "KSP install to use")]
+        public string KSP { get; set; }
+
+        [Option("kspdir", DefaultValue = null, HelpText = "KSP dir to use")]
+        public string KSPdir { get; set; }
+    }
+
+    // Each action defines its own options that it supports.
+    // Don't forget to cast to this type when you're processing them later on.
+
+    internal class InstallOptions : CommonOptions
+    {
+        [Option('c', "ckanfile", HelpText = "Local CKAN file to process")]
+        public string ckan_file { get; set; }
+
+        [Option("no-recommends", HelpText = "Do not install recommended modules")]
+        public bool no_recommends { get; set; }
+
+        [Option("with-suggests", HelpText = "Install suggested modules")]
+        public bool with_suggests { get; set; }
+
+        [Option("with-all-suggests", HelpText = "Install suggested modules all the way down")]
+        public bool with_all_suggests { get; set; }
+
+        // TODO: How do we provide helptext on this?
+        [ValueList(typeof (List<string>))]
+        public List<string> modules { get; set; }
+    }
+
+    internal class UpgradeOptions : CommonOptions
+    {
+        [Option('c', "ckanfile", HelpText = "Local CKAN file to process")]
+        public string ckan_file { get; set; }
+
+        [Option("no-recommends", HelpText = "Do not install recommended modules")]
+        public bool no_recommends { get; set; }
+
+        [Option("with-suggests", HelpText = "Install suggested modules")]
+        public bool with_suggests { get; set; }
+
+        [Option("with-all-suggests", HelpText = "Install suggested modules all the way down")]
+        public bool with_all_suggests { get; set; }
+
+        // TODO: How do we provide helptext on this?
+        [ValueList(typeof (List<string>))]
+        public List<string> modules { get; set; }
+    }
+
+    internal class ScanOptions : CommonOptions
+    {
+    }
+
+    internal class ListOptions : CommonOptions
+    {
+    }
+
+    internal class VersionOptions : CommonOptions
+    {
+    }
+
+    internal class CleanOptions : CommonOptions
+    {
+    }
+
+    internal class AvailableOptions : CommonOptions
+    {
+    }
+
+    internal class GuiOptions : CommonOptions
+    {
+    }
+
+    internal class UpdateOptions : CommonOptions
+    {
+        // This option is really meant for devs testing their CKAN-meta forks.
+        [Option('r', "repo", HelpText = "CKAN repository to use (experimental!)")]
+        public string repo { get; set; }
+    }
+
+    internal class RemoveOptions : CommonOptions
+    {
+        [ValueOption(0)]
+        public string Modname { get; set; }
+    }
+
+    internal class ShowOptions : CommonOptions
+    {
+        [ValueOption(0)]
+        public string Modname { get; set; }
+    }
+
+    internal class ListInstallsOptions : CommonOptions
+    {
+    }
+
+    internal class AddInstallOptions : CommonOptions
+    {
+        [ValueOption(0)]
+        public string name { get; set; }
+
+        [ValueOption(1)]
+        public string path { get; set; }
+    }
+
+    internal class RenameInstallOptions : CommonOptions
+    {
+        [ValueOption(0)]
+        public string old_name { get; set; }
+
+        [ValueOption(1)]
+        public string new_name { get; set; }
+    }
+
+    internal class RemoveInstallOptions : CommonOptions
+    {
+        [ValueOption(0)]
+        public string name { get; set; }
+    }
+
+    internal class SetDefaultInstallOptions : CommonOptions
+    {
+        [ValueOption(0)]
+        public string name { get; set; }
+    }
+
+    internal class ClearCacheOptions : CommonOptions
+    {
+    }
+}
+
